requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
google-genai==1.52.0
youtube-transcript-api==1.2.3
python-dotenv==1.2.1
yt-dlp==2025.11.12
